requirements.txt
# Core Flask dependencies
Flask==2.3.3
Flask-SocketIO==5.3.6
Flask-Compress==1.14

# CinePi integration
PyYAML==6.0.1
opencv-python==4.8.1.78

# System monitoring
psutil==5.9.6

# WebSocket and real-time
python-socketio==5.9.0
eventlet==0.33.3

# Development and testing
pytest==7.4.2
pytest-flask==1.3.0
black==23.9.1
flake8==6.1.0

# Production
gunicorn==21.2.0
